requirements.txt
Django==1.11
django-taggit==0.22.2
pytz==2017.3
